+++ src/lib/dashboard/product_dictionary.ts
@@ -1,0 +1,50 @@
+export const PRODUCTS = [
+  {
+    uuid: "2504d79b-c2ce-4660-9ea7-5576c8bb755f",
+    brand: "Rosumeg Combigel",
+    inn: "Rosuvastatin + Omega-3",
+    priority: true,
+  },
+  {
+    uuid: "859e60f9-8544-43b3-a6a0-f6c7529847eb",
+    brand: "Atmeg Combigel",
+    inn: "Atorvastatin + Omega-3",
+    priority: false,
+  },
+  {
+    uuid: "fcae4399-aa80-4318-ad55-89d6401c10a9",
+    brand: "Ciloduo",
+    inn: "Cilostazol + Rosuvastatin",
+    priority: false,
+  },
+  {
+    uuid: "24738c3b-3a5b-40a9-9e8e-889ec075b453",
+    brand: "Gastiin CR",
+    inn: "Mosapride 15mg",
+    priority: false,
+  },
+  {
+    uuid: "f88b87b8-c0ab-4f6e-ba34-e9330d1d4e18",
+    brand: "Omethyl Cutielet",
+    inn: "Omega-3 EE 2g",
+    priority: true,
+  },
+  {
+    uuid: "014fd4d2-dc66-4fc1-8d4f-59695183387f",
+    brand: "Sereterol Activair",
+    inn: "Salmeterol + Fluticasone",
+    priority: false,
+  },
+  {
+    uuid: "895f49ae-6ce3-44a3-93bd-bb77e027ba59",
+    brand: "Gadvoa Inj.",
+    inn: "Gadobutrol 604mg",
+    priority: false,
+  },
+  {
+    uuid: "bdfc9883-6040-438a-8e7a-df01f1230682",
+    brand: "Hydrine",
+    inn: "Hydroxyurea 500mg",
+    priority: false,
+  },
+] as const;